FBP-206. Moved class to correct package, added some basic logging
