+++ activiti-intellij-plugin/src/main/kotlin/com/valb3r/bpmn/intellij/activiti/plugin/ActivitiBpmnPluginToolWindowFactory.kt
@@ -1,23 +1,27 @@
 package com.valb3r.bpmn.intellij.activiti.plugin
 
 import com.intellij.openapi.components.ServiceManager
+import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory
 import com.valb3r.bpmn.intellij.activiti.plugin.popupmenu.ActivitiCanvasPopupMenuProvider
 import com.valb3r.bpmn.intellij.activiti.plugin.xmlnav.ActivitiXmlNavigator
-import com.valb3r.bpmn.intellij.plugin.BpmnPluginToolWindow
 import com.valb3r.bpmn.intellij.plugin.activiti.parser.ActivitiObjectFactory
 import com.valb3r.bpmn.intellij.plugin.activiti.parser.ActivitiParser
 import com.valb3r.bpmn.intellij.plugin.commons.langinjection.registerCurrentFile
+import com.valb3r.bpmn.intellij.plugin.core.BpmnPluginToolWindow
 import com.valb3r.bpmn.intellij.plugin.core.newelements.registerNewElementsFactory
 import com.valb3r.bpmn.intellij.plugin.core.ui.components.popupmenu.registerPopupMenuProvider
 import com.valb3r.bpmn.intellij.plugin.core.xmlnav.registerXmlNavigator
 
 class ActivitiBpmnPluginToolWindowFactory: ToolWindowFactory {
 
+    private val log = Logger.getInstance(ActivitiBpmnPluginToolWindowFactory::class.java)
+
     override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
+        log.info("Creating tool window content")
         registerPopupMenuProvider(ActivitiCanvasPopupMenuProvider())
         registerNewElementsFactory(ActivitiObjectFactory())
         val bpmnWindow = BpmnPluginToolWindow(ActivitiParser()) {
@@ -36,5 +40,6 @@ class ActivitiBpmnPluginToolWindowFactory: ToolWindowFactory {
                 false
         )
         toolWindow.contentManager.addContent(content)
+        log.info("Tool window content created")
     }
 }
