+++ activiti-intellij-plugin/src/main/kotlin/com/valb3r/bpmn/intellij/activiti/plugin/ActivitiBpmnPluginToolWindowProjectService.kt
@@ -1,6 +1,6 @@
 package com.valb3r.bpmn.intellij.activiti.plugin
 
-import com.valb3r.bpmn.intellij.plugin.BpmnPluginToolWindow
+import com.valb3r.bpmn.intellij.plugin.core.BpmnPluginToolWindow
 
 class ActivitiBpmnPluginToolWindowProjectService {
